detangle options
