+++ src/marks/cell.js
@@ -1,4 +1,4 @@
-import {identity, indexOf, maybeColorChannel, maybeTuple} from "../mark.js";
+import {identity, indexOf, maybeColorChannel, maybeTuple} from "../options.js";
 import {AbstractBar} from "./bar.js";
 
 export class Cell extends AbstractBar {
