+++ src/marks/frame.js
@@ -1,5 +1,6 @@
 import {create} from "d3";
-import {Mark, number} from "../mark.js";
+import {Mark} from "../mark.js";
+import {number} from "../options.js";
 import {applyDirectStyles, applyIndirectStyles, applyTransform, offset} from "../style.js";
 
 const defaults = {
