+++ src/options.js
@@ -1,13 +1,179 @@
+import {color} from "d3";
+
 // https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/TypedArray
 const TypedArray = Object.getPrototypeOf(Uint8Array);
 const objectToString = Object.prototype.toString;
 
-export function field(name) {
-  return d => d[name];
+export const field = name => d => d[name];
+export const constant = x => () => x;
+export const indexOf = (d, i) => i;
+export const identity = {transform: d => d};
+export const zero = () => 0;
+export const string = x => x == null ? x : `${x}`;
+export const number = x => x == null ? x : +x;
+export const boolean = x => x == null ? x : !!x;
+export const first = d => d[0];
+export const second = d => d[1];
+
+// A few extra color keywords not known to d3-color.
+const colors = new Set(["currentColor", "none"]);
+
+// Some channels may allow a string constant to be specified; to differentiate
+// string constants (e.g., "red") from named fields (e.g., "date"), this
+// function tests whether the given value is a CSS color string and returns a
+// tuple [channel, constant] where one of the two is undefined, and the other is
+// the given value. If you wish to reference a named field that is also a valid
+// CSS color, use an accessor (d => d.red) instead.
+export function maybeColorChannel(value, defaultValue) {
+  if (value === undefined) value = defaultValue;
+  return value === null ? [undefined, "none"]
+    : typeof value === "string" && (colors.has(value) || color(value)) ? [undefined, value]
+    : [value, undefined];
 }
 
-export function constant(value) {
-  return () => value;
+// Similar to maybeColorChannel, this tests whether the given value is a number
+// indicating a constant, and otherwise assumes that it’s a channel value.
+export function maybeNumberChannel(value, defaultValue) {
+  if (value === undefined) value = defaultValue;
+  return value === null || typeof value === "number" ? [undefined, value]
+    : [value, undefined];
+}
+
+// Validates the specified optional string against the allowed list of keywords.
+export function maybeKeyword(input, name, allowed) {
+  if (input != null) return keyword(input, name, allowed);
+}
+
+// Validates the specified required string against the allowed list of keywords.
+export function keyword(input, name, allowed) {
+  const i = `${input}`.toLowerCase();
+  if (!allowed.includes(i)) throw new Error(`invalid ${name}: ${input}`);
+  return i;
+}
+
+// For marks specified either as [0, x] or [x1, x2], such as areas and bars.
+export function maybeZero(x, x1, x2, x3 = identity) {
+  if (x1 === undefined && x2 === undefined) { // {x} or {}
+    x1 = 0, x2 = x === undefined ? x3 : x;
+  } else if (x1 === undefined) { // {x, x2} or {x2}
+    x1 = x === undefined ? 0 : x;
+  } else if (x2 === undefined) { // {x, x1} or {x1}
+    x2 = x === undefined ? 0 : x;
+  }
+  return [x1, x2];
+}
+
+// For marks that have x and y channels (e.g., cell, dot, line, text).
+export function maybeTuple(x, y) {
+  return x === undefined && y === undefined ? [first, second] : [x, y];
+}
+
+// A helper for extracting the z channel, if it is variable. Used by transforms
+// that require series, such as moving average and normalize.
+export function maybeZ({z, fill, stroke} = {}) {
+  if (z === undefined) ([z] = maybeColorChannel(fill));
+  if (z === undefined) ([z] = maybeColorChannel(stroke));
+  return z;
+}
+
+// Returns a Uint32Array with elements [0, 1, 2, … data.length - 1].
+export function range(data) {
+  return Uint32Array.from(data, indexOf);
+}
+
+// Returns a filtered range of data given the test function.
+export function where(data, test) {
+  return range(data).filter(i => test(data[i], i, data));
+}
+
+// Returns an array [values[index[0]], values[index[1]], …].
+export function take(values, index) {
+  return Array.from(index, i => values[i]);
+}
+
+export function maybeInput(key, options) {
+  if (options[key] !== undefined) return options[key];
+  switch (key) {
+    case "x1": case "x2": key = "x"; break;
+    case "y1": case "y2": key = "y"; break;
+  }
+  return options[key];
+}
+
+// Defines a channel whose values are lazily populated by calling the returned
+// setter. If the given source is labeled, the label is propagated to the
+// returned channel definition.
+export function lazyChannel(source) {
+  let value;
+  return [
+    {
+      transform: () => value,
+      label: labelof(source)
+    },
+    v => value = v
+  ];
+}
+
+export function labelof(value, defaultValue) {
+  return typeof value === "string" ? value
+    : value && value.label !== undefined ? value.label
+    : defaultValue;
+}
+
+// Like lazyChannel, but allows the source to be null.
+export function maybeLazyChannel(source) {
+  return source == null ? [source] : lazyChannel(source);
+}
+
+// Assuming that both x1 and x2 and lazy channels (per above), this derives a
+// new a channel that’s the average of the two, and which inherits the channel
+// label (if any). Both input channels are assumed to be quantitative. If either
+// channel is temporal, the returned channel is also temporal.
+export function mid(x1, x2) {
+  return {
+    transform(data) {
+      const X1 = x1.transform(data);
+      const X2 = x2.transform(data);
+      return isTemporal(X1) || isTemporal(X2)
+        ? Array.from(X1, (_, i) => new Date((+X1[i] + +X2[i]) / 2))
+        : Float64Array.from(X1, (_, i) => (+X1[i] + +X2[i]) / 2);
+    },
+    label: x1.label
+  };
+}
+
+// This distinguishes between per-dimension options and a standalone value.
+export function maybeValue(value) {
+  return value === undefined || isOptions(value) ? value : {value};
+}
+
+export function numberChannel(source) {
+  return {
+    transform: data => valueof(data, source, Float64Array),
+    label: labelof(source)
+  };
+}
+
+export function isOrdinal(values) {
+  for (const value of values) {
+    if (value == null) continue;
+    const type = typeof value;
+    return type === "string" || type === "boolean";
+  }
+}
+
+export function isTemporal(values) {
+  for (const value of values) {
+    if (value == null) continue;
+    return value instanceof Date;
+  }
+}
+
+export function isNumeric(values) {
+  for (const value of values) {
+    if (value == null) continue;
+    return typeof value === "number";
+  }
 }
 
 // This allows transforms to behave equivalently to channels.
