+++ src/scales.js
@@ -1,11 +1,11 @@
 import {descending} from "d3";
 import {parse as isoParse} from "isoformat";
+import {isOrdinal, isTemporal} from "./options.js";
 import {registry, color, position, radius, opacity, symbol, length} from "./scales/index.js";
 import {ScaleLinear, ScaleSqrt, ScalePow, ScaleLog, ScaleSymlog, ScaleQuantile, ScaleThreshold, ScaleIdentity} from "./scales/quantitative.js";
 import {ScaleDiverging, ScaleDivergingSqrt, ScaleDivergingPow, ScaleDivergingLog, ScaleDivergingSymlog} from "./scales/diverging.js";
 import {ScaleTime, ScaleUtc} from "./scales/temporal.js";
 import {ScaleOrdinal, ScalePoint, ScaleBand} from "./scales/ordinal.js";
-import {isOrdinal, isTemporal} from "./mark.js";
 
 export function Scales(channels, {
   inset: globalInset = 0,
