+++ src/transforms/interval.js
@@ -1,5 +1,4 @@
-import {labelof, maybeValue} from "../mark.js";
-import {valueof} from "../options.js";
+import {labelof, maybeValue, valueof} from "../options.js";
 import {maybeInsetX, maybeInsetY} from "./inset.js";
 
 // TODO Allow the interval to be specified as a string, e.g. “day” or “hour”?
